Sol: Fix WorldSession::HandlePetActionHelper
Spell ID has to be uint32.

+++ src/server/game/Handlers/PetHandler.cpp
@@ -475,7 +475,7 @@ void WorldSession::HandlePetStopAttack(WorldPacket &recvData)
     pet->ClearInPetCombat();
 }
 
-void WorldSession::HandlePetActionHelper(Unit* pet, uint64 guid1, uint16 spellid, uint16 flag, uint64 guid2)
+void WorldSession::HandlePetActionHelper(Unit* pet, uint64 guid1, uint32 spellid, uint16 flag, uint64 guid2)
 {
     CharmInfo* charmInfo = pet->GetCharmInfo();
     if (!charmInfo)
